+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 
 ### Fix
 
+* Ensure the connector is more robust in attempting to reconnect on losing connection. [Ben Dalling]
+
 * Migrate to the Qpid JMS SDK as it supports ASB and ActiveMQ/Artemis as an emulator. [Ben Dalling]
 
 ### Other